Move dataclass_transform code into prefab.py
